Moving PHPSpec to separate namespace
