+++ spec/PhpSpec/Definition/PHP/Methods/InitializableMethodSpec.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace spec\NullDev\Skeleton\Definition\PHP\Methods\PhpSpec;
+namespace spec\NullDev\Skeleton\PhpSpec\Definition\PHP\Methods;
 
 use NullDev\Skeleton\Definition\PHP\Methods\Method;
-use NullDev\Skeleton\Definition\PHP\Methods\PhpSpec\InitializableMethod;
+use NullDev\Skeleton\PhpSpec\Definition\PHP\Methods\InitializableMethod;
 use PhpSpec\ObjectBehavior;
 
 class InitializableMethodSpec extends ObjectBehavior
