+++ src/PhpSpec/CodeGenerator/PhpParser/Methods/ExposeConstructorArgumentsAsGettersGenerator.php
@@ -2,9 +2,8 @@
 
 declare(strict_types=1);
 
-namespace NullDev\Skeleton\CodeGenerator\PhpParser\Methods\PhpSpec;
+namespace NullDev\Skeleton\PhpSpec\CodeGenerator\PhpParser\Methods;
 
-use NullDev\Skeleton\Definition\PHP\Methods\PhpSpec\LetMethod;
 use NullDev\Skeleton\Definition\PHP\Parameter;
 use NullDev\Skeleton\Definition\PHP\Types\TypeDeclaration\ArrayType;
 use NullDev\Skeleton\Definition\PHP\Types\TypeDeclaration\BoolType;
@@ -12,9 +11,9 @@ use NullDev\Skeleton\Definition\PHP\Types\TypeDeclaration\FloatType;
 use NullDev\Skeleton\Definition\PHP\Types\TypeDeclaration\IntType;
 use NullDev\Skeleton\Definition\PHP\Types\TypeDeclaration\StringType;
 use NullDev\Skeleton\Definition\PHP\Types\TypeDeclaration\TypeDeclaration;
+use NullDev\Skeleton\PhpSpec\Definition\PHP\Methods\ExposeConstructorArgumentsAsGettersMethod;
 use PhpParser\BuilderFactory;
 use PhpParser\Node\Expr\Array_;
-use PhpParser\Node\Expr\Assign;
 use PhpParser\Node\Expr\MethodCall;
 use PhpParser\Node\Expr\Variable;
 use PhpParser\Node\Scalar\DNumber;
@@ -22,9 +21,10 @@ use PhpParser\Node\Scalar\LNumber;
 use PhpParser\Node\Scalar\String_;
 
 /**
+ * @see ExposeConstructorArgumentsAsGettersGeneratorSpec
  * @SuppressWarnings(PHPMD.CouplingBetweenObjects)
  */
-class LetGenerator
+class ExposeConstructorArgumentsAsGettersGenerator
 {
     private $builderFactory;
 
@@ -33,52 +33,69 @@ class LetGenerator
         $this->builderFactory = $builderFactory;
     }
 
-    public function generate(LetMethod $method)
+    public function generate(ExposeConstructorArgumentsAsGettersMethod $method)
     {
         $node = $this->builderFactory
             ->method($method->getMethodName())
             ->makePublic();
 
-        $beParams = [];
-
         foreach ($method->getMethodParameters() as $param) {
-            if ($this->isParameterEligibleForLetParameter($param)) {
+            if ($this->isParameterEligibleForMethodParameter($param)) {
                 $node->addParam($this->createMethodParam($param));
-                $beParams[] = new Variable($param->getName());
-            } else {
-                $beParams[] = $this->createBeConstructedWithArgument($param);
             }
         }
 
-        $node->addStmt(new MethodCall(new Variable('this'), 'beConstructedWith', $beParams));
+        foreach ($method->getMethodParameters() as $param) {
+            if ($this->isParameterEligibleForMethodParameter($param)) {
+                $node->addStmt(
+                    new MethodCall(
+                        new MethodCall(
+                            new Variable('this'),
+                            'get'.ucfirst($param->getName())
+                        ),
+                        'shouldReturn',
+                        [new Variable($param->getName())]
+                    )
+                );
+            } else {
+                $node->addStmt(
+                    new MethodCall(
+                        new MethodCall(
+                            new Variable('this'),
+                            'get'.ucfirst($param->getName())
+                        ),
+                        'shouldReturn',
+                        [$this->createShouldReturn($param)]
+                    )
+                );
+            }
+        }
 
         return $node;
     }
 
-    private function createBeConstructedWithArgument(Parameter $parameter)
+    private function createShouldReturn(Parameter $parameter)
     {
-        $variable = new Variable($parameter->getName());
-
         if (false === $parameter->hasClass()) {
-            return new Assign($variable, new String_($parameter->getName()));
+            return new String_($parameter->getName());
         }
 
         if ($parameter->getClassType() instanceof StringType) {
-            return new Assign($variable, new String_($parameter->getName()));
+            return new String_($parameter->getName());
         } elseif ($parameter->getClassType() instanceof IntType) {
-            return new Assign($variable, new LNumber(1));
+            return new LNumber(1);
         } elseif ($parameter->getClassType() instanceof ArrayType) {
-            return new Assign($variable, new Array_([], ['kind' => Array_::KIND_SHORT]));
+            return new Array_([], ['kind' => Array_::KIND_SHORT]);
         } elseif ($parameter->getClassType() instanceof FloatType) {
-            return new Assign($variable, new DNumber(2.0));
+            return new DNumber(2.0);
         } elseif ($parameter->getClassType() instanceof BoolType) {
-            return new Assign($variable, true);
+            return true;
         }
 
-        throw new \Exception('ERR 90131234: Unhandled argument received.');
+        throw new \Exception('ERR 242342123123: Unhandled argument received.');
     }
 
-    private function isParameterEligibleForLetParameter(Parameter $parameter)
+    private function isParameterEligibleForMethodParameter(Parameter $parameter)
     {
         if (false === $parameter->hasClass()) {
             return false;
@@ -94,10 +111,6 @@ class LetGenerator
     private function createMethodParam(Parameter $param)
     {
         $result = $this->builderFactory->param($param->getName());
-
-        if ($param->hasClass()) {
-            $result->setTypeHint($param->getClassType()->getName());
-        }
 
         return $result;
     }
